requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-nested-routers==0.93.3
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5